+++ src/test/kotlin/com/hhplus/concert/application/facade/unittest/QueueServiceTest.kt
@@ -1,4 +1,4 @@
-package com.hhplus.concert.application.facade
+package com.hhplus.concert.application.facade.unittest
 
 import com.hhplus.concert.business.application.service.QueueService
 import com.hhplus.concert.business.domain.manager.queue.QueueManager
