+++ src/test/kotlin/com/hhplus/concert/application/facade/unittest/ReservationServiceTest.kt
@@ -1,4 +1,4 @@
-package com.hhplus.concert.application.facade
+package com.hhplus.concert.application.facade.unittest
 
 import com.hhplus.concert.business.application.dto.ReservationServiceDto
 import com.hhplus.concert.business.application.service.ReservationService
